Fix: components moved to view directory
