+++ src/main.js
@@ -1,11 +1,11 @@
-import {createSiteMenuTemplate} from './createSiteMenuTemplate';
-import {createTripInfoTemplate} from './createTripInfoTemplate';
-import {createTripFilterTemplate} from './createTripFilterTemplate';
-import {createTripSortTemplate} from './createTripSortTemplate';
-import {createTripEventsListTemplate} from './createTripEventsListTemplate';
-import {createAddNewPointFormTemplate} from './createAddNewPointFormTemplate';
-import {createEventTemplate} from './createEventTemplate';
-import {createEditPointForm} from './createEditPointForm';
+import {createSiteMenuTemplate} from './view/createSiteMenuTemplate';
+import {createTripInfoTemplate} from './view/createTripInfoTemplate';
+import {createTripFilterTemplate} from './view/createTripFilterTemplate';
+import {createTripSortTemplate} from './view/createTripSortTemplate';
+import {createTripEventsListTemplate} from './view/createTripEventsListTemplate';
+import {createAddNewPointFormTemplate} from './view/createAddNewPointFormTemplate';
+import {createEventTemplate} from './view/createEventTemplate';
+import {createEditPointForm} from './view/createEditPointForm';
 
 const headerMain = document.querySelector('.trip-main');
 const pageMain = document.querySelector('.page-main');
