+++ src/view/createAddNewPointFormTemplate.js
