+++ src/view/createEventTemplate.js
