+++ src/view/createTripEventsItemTemplate.js
