+++ src/view/createTripEventsListTemplate.js
