+++ src/view/createTripInfoTemplate.js
