+++ src/view/createTripSortTemplate.js
